Clean TLS directory
